src/requirements_minus_torch.txt
datasets==2.13.1
transformers==4.30.2
Pillow==10.0.0
pytorch-lightning==1.9.0
bs4==0.0.1
nltk==3.8.1
PyMuPDF==1.22.5
configargparse==1.5.5
lxml==4.9.3
matplotlib==3.7.2
tensorboard==2.13.0